Refactoring архитектуры моделей, с учётом принципа единой ответственности.
