+++ models/call/CallActiveRecord.php
@@ -1,8 +1,10 @@
 <?php
 declare(strict_types = 1);
 
-namespace app\models;
+namespace app\models\call;
 
+use app\models\customer\Customer;
+use app\models\user\User;
 use Yii;
 use yii\db\ActiveQuery;
 use yii\db\ActiveRecord;
@@ -32,7 +34,7 @@ use yii\db\ActiveRecord;
  * @property Customer $customer
  * @property User $user
  */
-class Call extends ActiveRecord
+class CallActiveRecord extends ActiveRecord
 {
     public const STATUS_NO_ANSWERED = 0;
     public const STATUS_ANSWERED = 1;
@@ -112,84 +114,5 @@ class Call extends ActiveRecord
     public function getUser(): ActiveQuery
     {
         return $this->hasOne(User::class, ['id' => 'user_id']);
-    }
-
-    /**
-     * @return string
-     */
-    public function getClient_phone(): string
-    {
-        return $this->direction === self::DIRECTION_INCOMING ? $this->phone_from : $this->phone_to;
-    }
-
-    /**
-     * @return mixed|string
-     */
-    public function getTotalStatusText(): string
-    {
-        if (
-            $this->status === self::STATUS_NO_ANSWERED
-            && $this->direction === self::DIRECTION_INCOMING
-        ) {
-            return Yii::t('app', 'Missed Call');
-        }
-
-        if (
-            $this->status === self::STATUS_NO_ANSWERED
-            && $this->direction === self::DIRECTION_OUTGOING
-        ) {
-            return Yii::t('app', 'Client No Answer');
-        }
-
-        $msg = $this->getFullDirectionText();
-
-        if ($this->duration) {
-            $msg .= ' (' . $this->getDurationText() . ')';
-        }
-
-        return $msg;
-    }
-
-    /**
-     * @param bool $hasComment
-     * @return string
-     */
-    public function getTotalDisposition(bool $hasComment = true): string
-    {
-        $t = [];
-        if ($hasComment && $this->comment) {
-            $t[] = $this->comment;
-        }
-        return implode(': ', $t);
-    }
-
-    /**
-     * @return array
-     */
-    public static function getFullDirectionTexts(): array
-    {
-        return [
-            self::DIRECTION_INCOMING => Yii::t('app', 'Incoming Call'),
-            self::DIRECTION_OUTGOING => Yii::t('app', 'Outgoing Call'),
-        ];
-    }
-
-    /**
-     * @return mixed|string
-     */
-    public function getFullDirectionText()
-    {
-        return self::getFullDirectionTexts()[$this->direction] ?? $this->direction;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDurationText(): string
-    {
-        if (null !== $this->duration) {
-            return $this->duration >= 3600 ? gmdate("H:i:s", $this->duration) : gmdate("i:s", $this->duration);
-        }
-        return '00:00';
     }
 }
