+++ models/fax/FaxActiveRecord.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace app\models;
+namespace app\models\fax;
 
+use app\models\user\User;
 use Yii;
 use yii\db\ActiveQuery;
 use yii\db\ActiveRecord;
@@ -22,14 +23,8 @@ use yii\db\ActiveRecord;
  *
  * @property User $user
  */
-class Fax extends ActiveRecord
+class FaxActiveRecord extends ActiveRecord
 {
-    public const DIRECTION_INCOMING = 0;
-    public const DIRECTION_OUTGOING = 1;
-
-    public const TYPE_POA_ATC = 'poa_atc';
-    public const TYPE_REVOCATION_NOTICE = 'revocation_notice';
-
     /**
      * @inheritdoc
      */
@@ -79,24 +74,4 @@ class Fax extends ActiveRecord
     {
         return $this->hasOne(User::class, ['id' => 'user_id']);
     }
-
-    /**
-     * @return array
-     */
-    public static function getTypeTexts(): array
-    {
-        return [
-            self::TYPE_POA_ATC => Yii::t('app', 'POA/ATC'),
-            self::TYPE_REVOCATION_NOTICE => Yii::t('app', 'Revocation'),
-        ];
-    }
-
-    /**
-     * @return mixed|string
-     */
-    public function getTypeText()
-    {
-        return self::getTypeTexts()[$this->type] ?? $this->type;
-    }
-
 }
