+++ models/history/History.php
@@ -1,39 +1,15 @@
 <?php
 declare(strict_types = 1);
 
-namespace app\models;
+namespace app\models\history;
 
-use app\models\traits\ObjectNameTrait;
 use Yii;
-use yii\db\ActiveQuery;
-use yii\db\ActiveRecord;
 
 /**
- * This is the model class for table "{{%history}}".
- *
- * @property int $id
- * @property string $ins_ts
- * @property int $customer_id
- * @property string $event
- * @property string $object
- * @property int $object_id
- * @property string $message
- * @property string $detail
- * @property int $user_id
- *
- * @property string $eventText
- *
- * @property Customer $customer
- * @property User $user
- *
- * @property Task $task
- * @property Sms $sms
- * @property Call $call
+ * Бизнес логика модели историй.
  */
-class History extends ActiveRecord
+class History extends HistoryActiveRecord
 {
-    use ObjectNameTrait;
-
     public const EVENT_CREATED_TASK = 'created_task';
     public const EVENT_UPDATED_TASK = 'updated_task';
     public const EVENT_COMPLETED_TASK = 'completed_task';
@@ -49,76 +25,6 @@ class History extends ActiveRecord
 
     public const EVENT_CUSTOMER_CHANGE_TYPE = 'customer_change_type';
     public const EVENT_CUSTOMER_CHANGE_QUALITY = 'customer_change_quality';
-
-    /**
-     * @inheritdoc
-     */
-    public static function tableName(): string
-    {
-        return '{{%history}}';
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function rules()
-    {
-        return [
-            [['ins_ts'], 'safe'],
-            [['customer_id', 'object_id', 'user_id'], 'integer'],
-            [['event'], 'required'],
-            [['message', 'detail'], 'string'],
-            [['event', 'object'], 'string', 'max' => 255],
-            [
-                ['customer_id'],
-                'exist',
-                'skipOnError' => true,
-                'targetClass' => Customer::class,
-                'targetAttribute' => ['customer_id' => 'id']
-            ],
-            [
-                ['user_id'],
-                'exist',
-                'skipOnError' => true,
-                'targetClass' => User::class,
-                'targetAttribute' => ['user_id' => 'id']
-            ],
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels(): array
-    {
-        return [
-            'id' => Yii::t('app', 'ID'),
-            'ins_ts' => Yii::t('app', 'Ins Ts'),
-            'customer_id' => Yii::t('app', 'Customer ID'),
-            'event' => Yii::t('app', 'Event'),
-            'object' => Yii::t('app', 'Object'),
-            'object_id' => Yii::t('app', 'Object ID'),
-            'message' => Yii::t('app', 'Message'),
-            'detail' => Yii::t('app', 'Detail'),
-            'user_id' => Yii::t('app', 'User ID'),
-        ];
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getCustomer(): ActiveQuery
-    {
-        return $this->hasOne(Customer::class, ['id' => 'customer_id']);
-    }
-
-    /**
-     * @return ActiveQuery
-     */
-    public function getUser(): ActiveQuery
-    {
-        return $this->hasOne(User::class, ['id' => 'user_id']);
-    }
 
     /**
      * @return array
