+++ models/task/TaskActiveRecord.php
@@ -1,8 +1,10 @@
 <?php
 declare(strict_types = 1);
 
-namespace app\models;
+namespace app\models\task;
 
+use app\models\customer\Customer;
+use app\models\user\User;
 use Yii;
 use yii\db\ActiveQuery;
 use yii\db\ActiveRecord;
@@ -34,16 +36,8 @@ use yii\db\ActiveRecord;
  * @property string $isInbox
  * @property string $statusText
  */
-class Task extends ActiveRecord
+class TaskActiveRecord extends ActiveRecord
 {
-    public const STATUS_NEW = 0;
-    public const STATUS_DONE = 1;
-    public const STATUS_CANCEL = 3;
-
-    public const STATE_INBOX = 'inbox';
-    public const STATE_DONE = 'done';
-    public const STATE_FUTURE = 'future';
-
     /**
      * @inheritdoc
      */
@@ -112,71 +106,5 @@ class Task extends ActiveRecord
     public function getUser(): ActiveQuery
     {
         return $this->hasOne(User::class, ['id' => 'user_id']);
-    }
-
-    /**
-     * @return array
-     */
-    public static function getStatusTexts(): array
-    {
-        return [
-            self::STATUS_NEW => Yii::t('app', 'New'),
-            self::STATUS_DONE => Yii::t('app', 'Complete'),
-            self::STATUS_CANCEL => Yii::t('app', 'Cancel'),
-        ];
-    }
-
-    /**
-     * @param $value
-     * @return int|mixed
-     */
-    public function getStatusTextByValue($value): int
-    {
-        return self::getStatusTexts()[$value] ?? $value;
-    }
-
-    /**
-     * @return mixed|string
-     */
-    public function getStatusText(): string
-    {
-        return $this->getStatusTextByValue($this->status);
-    }
-
-    /**
-     * @return array
-     */
-    public static function getStateTexts(): array
-    {
-        return [
-            self::STATE_INBOX => Yii::t('app', 'Inbox'),
-            self::STATE_DONE => Yii::t('app', 'Done'),
-            self::STATE_FUTURE => Yii::t('app', 'Future')
-        ];
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getStateText()
-    {
-        return self::getStateTexts()[$this->state] ?? $this->state;
-    }
-
-
-    /**
-     * @return bool
-     */
-    public function getIsOverdue(): bool
-    {
-        return $this->status !== self::STATUS_DONE && strtotime($this->due_date) < time();
-    }
-
-    /**
-     * @return bool
-     */
-    public function getIsDone(): bool
-    {
-        return $this->status === self::STATUS_DONE;
     }
 }
